refactor: new filters and list component seem to be working
